sync source (#2)
* Revert "msm: vidc: add ion_handle checking before mapping buffers."

This reverts commit 2e0150f4b314d018a2a903530086df666c65073a.

* msm: vidc: Fix SYS_ERROR handling

Do not clean the session while handling sys error, instead
only set the instance state to invalid and report the error to client.
The session will be cleaned as part of vidc_close.

CRs-Fixed: 755740
Change-Id: Iaa1bc9a39c51be67649830bf9cb6acdbce105211
Signed-off-by: Sowmya Pandiri <spandiri@codeaurora.org>

* msm: vidc: Handle max clients error properly

When FW returned max clients error, driver is changing inst->state
to INVALID. With this driver doesn't know if FW had valid session or
not to send ABORT command. If driver sends ABORT command, FW times out
because there is no session in FW. Hence clean the session in HFI when
FW returns max clients error so that ABORT doesn't reach FW.

Change-Id: I4da5b457b30415458ce79cf7b0de0bf22f82f8e3
Signed-off-by: Praneeth Paladugu <ppaladug@codeaurora.org>

* msm: vidc: fix null pointer crash in sys error handler

When the iteration pointer of list_for_each_entry is
used to evaluate, it can have a garbage value. So,instead,
using a inst pointer, using core to get the hdev.

Also, flushing the core work queue in unload firmware
leads to deadlock, when there is a SYS error is received
from core, because handle_sys_error is also a part of core
work handler, so flushing itself by its own execution,
will lead to wait indefinitely, i.e, a deadlock. So,
skipping it when there is a error condition.

Change-Id: Ic4e2df0db2341c974cbba19ddd55ecdc5c832124
Signed-off-by: Balamurugan Alagarsamy <balaga@codeaurora.org>

* msm: vidc: use dedicated pendingq.lock for pendingq

when the release buffer reference event is received
the sync_lock is acquired in handle_event_change()
and if the instance state is not START_DONE then
the same thread tries to acquire the sync_lock again
in msm_comm_qbuf() leading to deadlock issues.
So use dedicated pending queue lock instead of sync_lock
for pending queue buffer list to avoid the issues
due to sync_lock.

CRs-Fixed: 931529
Change-Id: I1037f5e0671dfebd8fe0701b5539908daecb1b5f
Signed-off-by: Balamurugan Alagarsamy <balaga@codeaurora.org>

* msm: vidc: Fix timestamp for EOS buffer

Timestamp for zero filled length FBD is not
being set correctly even if the timestamp from FW
is valid. This causes certain decode usecase
to not exit correctly.

CRs-fixed: 664086
Change-Id: Ieb0f274efa5d94c9294b021e061bca6d9b45de87
Signed-off-by: Sachin Shah <sachins@codeaurora.org>

* msm: vidc: initialize the session list when device initialized

When multiple sessions are running and if there is a SYS_ERROR
all clients will be notified with hardware error and clients
are expected to close the sessions. If new session is opened
before all sessions are closed results in list corruption
because session open will do list initialization. To avoid
list corruption issue initialize the session list at the time of
device initialization instead of first session open.

Change-Id: Ifd5c62ff7ae0547ecf73029fadbcac66f1e189bc
Signed-off-by: Balamurugan Alagarsamy <balaga@codeaurora.org>

* msm: vidc: Fix a race condition in error handling

Currently video driver handles session init and clean in one
thread and response handling from HW in another thread. Due to
race conditions response handler thread may access the session
which was already freed by forward thread. Hence synchronize the
forward thread and response handler with session_lock.

Ex : When driver detects SYS_ERROR, it informs client and client
tries to clean the session. In the meantime if HW sends a response,
the response handler thread tries to access the inst struct and this
struct might have freed as part of session close. This causes
use-after-free exception.

CRs-Fixed: 765521
Change-Id: I11519087e9c0dcb95587c6a048853e47032bd1be
Signed-off-by: Balamurugan Alagarsamy <balaga@codeaurora.org>

* msm: vidc: skip fps round off when client sets 1 fps

sometimes fps is rounded off one lesser, makes it
to zero, when the client is setting fps as 1.
This exposes division by zero crash when calculating
the load for non realtime usescases. Crash is observed
in recent monkey runs.

Change-Id: I4ec824d658ed3a14ab43ca031e989b01adfddc11
CRs-Fixed: 933641
Signed-off-by: Balamurugan Alagarsamy <balaga@codeaurora.org>

* msm: vidc: Make buffer validity checks stronger

Check for the exact number of planes that we advertised to the client
rather than the worst-case checks.

Change-Id: Ibaf705367db98beb0e01bb2c3087126cf2ba73e8
Signed-off-by: Deva Ramasubramanian <dramasub@codeaurora.org>

* msm: vidc: Clean up video instance locking logic

Clean up unnecessary locks in video driver code.

Change-Id: Ib00527531863a06fa0bad28ed3cc71fbdf167e2d
Signed-off-by: Arun Menon <avmenon@codeaurora.org>
Signed-off-by: Vasantha Balla <vballa@codeaurora.org>

* msm: vidc: Use registeredbuf.lock for buffer synchronization

Right now dynamic buffers are managed with both sync_lock and
registeredbufs.lock. sync_lock is meant only for state
transition. Hence remove sync_lock and achieve the same purpose
with registeredbufs.lock.

Change-Id: I3b0eb62019e7e992f63c159bf65d8c21cf2f52c8
Signed-off-by: Praneeth Paladugu <ppaladug@codeaurora.org>
Signed-off-by: Vikash Garodia <vgarodia@codeaurora.org>
Signed-off-by: Vasantha Balla <vballa@codeaurora.org>

* msm: vidc: Add same fd ref only for static buffer mode

Add same fd ref only for static buffer mode.
Driver is configured for dynamic buffer mode by
default on Venus 3xx targets. If client uses the
same fd for all output buffers, then the first
buffer can remain mapped without reference, which
is incompatible with the curently logic which
assumes buffers to have a reference if they are
already mapped

CRs-Fixed: 773605
Change-Id: I025fe373532e185660d43bae28457cbf06f20e7a
Signed-off-by: Surajit Podder <spodder@codeaurora.org>

* msm: vidc: add ion_handle checking before mapping buffers.

Compare ion handles in driver instead of matching fds
to check if a buffer is already mapped or not .

Bug: 28747768

Change-Id: Ifd18d8689351c4a6a22c988d359fb413be19e142
Signed-off-by: Ashray Kulkarni <ashrayk@codeaurora.org>
Signed-off-by: Praveen Chavan <pchavan@codeaurora.org>
Signed-off-by: Deva Ramasubramanian <dramasub@codeaurora.org>
Signed-off-by: Arun Menon <avmenon@codeaurora.org>
Signed-off-by: Suman Mukherjee <sumam@codeaurora.org>
Signed-off-by: Paras Nagda <pnagda@codeaurora.org>
Signed-off-by: Vasantha Balla <vballa@codeaurora.org>